PI config to send SMS with template

+++ Arduino/LM35TempRead/LM35TempRead.ino
@@ -20,7 +20,7 @@ temperature = reading * 0.488;
 }
 
 void sendData(){
-    Wire.write(temperature);
+    Wire.write((int)temperature);
 }
